Layers/requirements.txt
# ============================================================
# LAMBDA LAYER - DEPENDENCIAS PYTHON COMPARTIDAS
# ============================================================

# AWS SDK (común en todos los servicios)
boto3==1.34.34

# Variables de entorno
python-dotenv==1.0.0

# JWT para autenticación (Usuarios)
PyJWT==2.8.0

# FastAPI para APIs (Ingesta)
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Pydantic para validación de datos (Ingesta)
pydantic==2.5.3
pydantic-settings==2.1.0